# docs/api/assignment.openapi.yaml
openapi: 3.0.1
info:
  description: Coursistant LMS REST API contract (OpenAPI 3).
  title: LMS API
  version: 1.0.0
servers:
- description: Servlet context-path
  url: /api
security:
- bearerAuth: []
...
paths:
  /v2/courses/{courseId}/assignments:
    get:
      description: Instructor/TA see all states including Draft. Students only see
        Published; drafts are omitted. Response shape is role-specific (staff counters
        vs student submission state).
      operationId: assignmentList
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: OK
      summary: List assignments in a course
      tags:
      - Assignments
    post:
      description: Instructor/TA with edit permission. dueAt/lateUntil are wall-clock
        times in the course tenant timezone. Group assignments require submissionType=Group
        and groupSetId.
      operationId: assignmentCreate
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
        required: true
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "PARAM_MISSING, BAD_REQUEST, ASSIGNMENT_FILE_CONSTRAINT_INVALID"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_GROUP_SET_REQUIRED
      summary: Create an assignment
      tags:
      - Assignments
  /v2/courses/{courseId}/assignments/summaries:
    get:
      description: Any course member. Students only see Published; submissionStatus
        is student-oriented.
      operationId: assignmentListSummaries
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: OK
      summary: List slim assignment summary cards
      tags:
      - Assignments
  /v2/courses/{courseId}/assignments/{assignmentId}:
    delete:
      description: Instructor/TA. May fail when submissions already exist.
      operationId: assignmentDelete
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_HAS_SUBMISSIONS
      summary: Delete an assignment
      tags:
      - Assignments
    get:
      description: Instructor/TA may open Draft. Students receive ASSIGNMENT_NOT_FOUND
        for non-Published assignments. Staff get roster/grading counters; students
        get submission/grade view fields.
      operationId: assignmentGet
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND (including student access to Draft) or
            COURSE_NOT_FOUND
      summary: Get assignment detail
      tags:
      - Assignments
    patch:
      description: Instructor/TA. Partial update; only non-null fields apply. Shortening
        due date may require confirmShortenDueDate=true.
      operationId: assignmentPatch
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAssignmentRequest"
        required: true
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "ASSIGNMENT_TYPE_LOCKED, ASSIGNMENT_GROUP_SET_REQUIRED, ASSIGNMENT_DUE_SHORTEN_CONFIRM_REQUIRED,\
            \ ASSIGNMENT_HAS_SUBMISSIONS"
      summary: Patch an assignment
      tags:
      - Assignments
  /v2/courses/{courseId}/assignments/{assignmentId}/attachments:
    post:
      description: Instructor/TA. Multipart part `files` (array of binary files).
      operationId: assignmentUploadAttachments
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      - in: query
        name: files
        required: false
        schema:
          type: array
          items:
            type: string
            format: binary
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
        required: true
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "UNSUPPORTED_FILE_TYPE, FILE_TOO_LARGE, PARAM_MISSING"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND
      summary: Upload instructor attachments
      tags:
      - Assignments
  /v2/courses/{courseId}/assignments/{assignmentId}/attachments/{attachmentId}:
    delete:
      description: Instructor/TA only.
      operationId: assignmentDeleteAttachment
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: attachmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND or attachment NOT_FOUND
      summary: Delete an instructor attachment
      tags:
      - Assignments
  /v2/courses/{courseId}/assignments/{assignmentId}/attachments/{attachmentId}/download:
    get:
      description: Course members with assignment visibility. Returns raw bytes (not
        JSON ApiResponse). Students cannot download attachments on Draft assignments
        (ASSIGNMENT_NOT_FOUND).
      operationId: assignmentDownloadAttachment
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: attachmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
          description: Attachment bytes
          headers:
            Content-Disposition:
              description: attachment; filename="..."; filename*=UTF-8''...
              schema:
                type: string
                example: attachment; filename="brief.pdf"; filename*=UTF-8''brief.pdf
              style: simple
            Content-Type:
              description: "Stored content type, or application/octet-stream when\
                \ unknown"
              schema:
                type: string
                example: application/pdf
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND or attachment NOT_FOUND
      summary: Download an assignment attachment
      tags:
      - Assignments
...
  /v2/courses/{courseId}/assignments/{assignmentId}/due-date-change-preview:
    post:
      description: Instructor/TA dry-run. Writes nothing; reports who would become
        late / need confirmation.
      operationId: assignmentPreviewDueDateChange
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DueDateChangePreviewRequest"
        required: true
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND
      summary: Preview due-date change impact
      tags:
      - Assignments
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      description: Error envelope rendered by ApiExceptionHandler (data optional).
      properties:
        code:
          type: string
          description: Stable ErrorType.name()
          example: USER_NOT_FOUND
        data:
          type: object
          nullable: true
        message:
          type: string
          example: User Does Not Exist
        status:
          type: integer
          format: int32
          example: 404
        timestamp:
          type: string
          format: date-time
          example: 2026-07-23T10:00:00Z
      required:
      - code
      - message
      - status
      - timestamp
    ApiResponse:
      type: object
      description: Unified JSON envelope for LMS APIs
      properties:
        code:
          type: string
          description: Stable machine code (ErrorType.name() or SUCCESS)
          example: SUCCESS
        data:
          type: array
          description: Payload; omitted or null on many errors
          items:
            $ref: "#/components/schemas/UpcomingAssignmentDeadlineResponse"
        message:
          type: string
          description: Human-readable message
          example: Success
        status:
          type: integer
          format: int32
          description: HTTP status mirrored in body
          example: 200
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp
          example: 2026-07-23T10:00:00Z
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Access token from POST /v1/auth/login (Authorization: Bearer <token>)."
      name: bearerAuth
      scheme: bearer
      type: http